Add files via upload
PROBLEM STATEMENT:
Currently in schools teachers perform manual work to prepare and track students' learning outcome through report cards. This is the standard procedure done manually throughout the country. Often it is observed that students are also good in co-curricular activities and being good in academics is not the only way ahead. In the current situation it is not possible for a teacher to pay attention to every student and find out the best in a child. Hence there is a need to develop solutions to track all activities of a student including the academic progress in real-time. Innovation in this domain will be appreciated to develop systems which automatically reads the students progress and generates feedback also is a good way to proceed.

I have used HTML,CSS for front end and JSP for server side programming and mysql for DB.

Server: Apache Tomcat 8.5

1.Start the server
2.Extract and load all the files in root folder available in web apps of apache folder
3.set up the mysql DB accordingly (skilltracker sql file is uploaded)
4.Navigate to http://localhost:8080/home.html
5.Do appropriate actions